+++ fe/fe-core/src/test/java/com/starrocks/analysis/RefreshMaterializedViewTest.java
@@ -24,6 +24,7 @@ import com.starrocks.catalog.Partition;
 import com.starrocks.catalog.Replica;
 import com.starrocks.catalog.Table;
 import com.starrocks.catalog.Tablet;
+import com.starrocks.clone.DynamicPartitionScheduler;
 import com.starrocks.common.Config;
 import com.starrocks.pseudocluster.PseudoCluster;
 import com.starrocks.qe.ConnectContext;
@@ -364,5 +365,87 @@ public class RefreshMaterializedViewTest {
         Assert.assertEquals(1, versionMap2.size());
         Set<String> partitions2 = versionMap2.values().iterator().next().keySet();
         Assert.assertEquals(3, partitions2.size());
+    }
+
+    @Test
+    public void testDropBaseTablePartitionRemoveVersionMap() throws Exception {
+        new MockUp<StmtExecutor>() {
+            @Mock
+            public void handleDMLStmt(ExecPlan execPlan, DmlStmt stmt) throws Exception {
+                if (stmt instanceof InsertStmt) {
+                    InsertStmt insertStmt = (InsertStmt) stmt;
+                    TableName tableName = insertStmt.getTableName();
+                    Database testDb = GlobalStateMgr.getCurrentState().getDb(stmt.getTableName().getDb());
+                    OlapTable tbl = ((OlapTable) testDb.getTable(tableName.getTbl()));
+                    for (Partition partition : tbl.getPartitions()) {
+                        if (insertStmt.getTargetPartitionIds().contains(partition.getId())) {
+                            long version = partition.getVisibleVersion() + 1;
+                            partition.setVisibleVersion(version, System.currentTimeMillis());
+                            MaterializedIndex baseIndex = partition.getBaseIndex();
+                            List<Tablet> tablets = baseIndex.getTablets();
+                            for (Tablet tablet : tablets) {
+                                List<Replica> replicas = ((LocalTablet) tablet).getImmutableReplicas();
+                                for (Replica replica : replicas) {
+                                    replica.updateVersionInfo(version, -1, version);
+                                }
+                            }
+                        }
+                    }
+                }
+            }
+        };
+
+        starRocksAssert.useDatabase("test")
+            .withTable("CREATE TABLE `test`.`tbl_with_partition` (\n" +
+                "  `k1` date,\n" +
+                "  `k2` int,\n" +
+                "  `v1` string\n" +
+                ") ENGINE=OLAP\n" +
+                "DUPLICATE KEY(`k1`, `k2`)\n" +
+                "PARTITION BY RANGE(`k1`)\n" +
+                "(\n" +
+                "PARTITION p20230410 VALUES [(\"2023-04-10\"), (\"2023-04-11\")),\n" +
+                "PARTITION p20230411 VALUES [(\"2023-04-11\"), (\"2023-04-12\")),\n" +
+                "PARTITION p20230412 VALUES [(\"2023-04-12\"), (\"2023-04-13\")),\n" +
+                "PARTITION p20230413 VALUES [(\"2023-04-13\"), (\"2023-04-14\"))\n" +
+                ")\n" +
+                "DISTRIBUTED BY HASH(`v1`) BUCKETS 3\n" +
+                "PROPERTIES (\"replication_num\" = \"1\");")
+            .withMaterializedView("CREATE MATERIALIZED VIEW `mv_with_partition`\n" +
+                "PARTITION BY (date_trunc('day', k1))\n" +
+                "REFRESH DEFERRED MANUAL \n" +
+                "DISTRIBUTED BY HASH(`k1`) BUCKETS 3\n" +
+                "PROPERTIES (\"replication_num\" = \"1\", \"partition_refresh_number\"=\"1\")\n" +
+                "AS\n" +
+                "SELECT \n" +
+                "k1,\n" +
+                "count(DISTINCT `v1`) AS `v` \n" +
+                "FROM `test`.`tbl_with_partition`\n" +
+                "group by k1;");
+        starRocksAssert.updateTablePartitionVersion("test", "tbl_with_partition", 2);
+        starRocksAssert.refreshMvPartition("REFRESH MATERIALIZED VIEW test.mv_with_partition \n" +
+            "PARTITION START (\"2023-04-10\") END (\"2023-04-14\");");
+        MaterializedView mv = getMv("test", "mv_with_partition");
+        Map<Long, Map<String, MaterializedView.BasePartitionInfo>> versionMap =
+            mv.getRefreshScheme().getAsyncRefreshContext().getBaseTableVisibleVersionMap();
+        Assert.assertEquals(1, versionMap.size());
+        Set<String> partitions = versionMap.values().iterator().next().keySet();
+        Assert.assertEquals(4, partitions.size());
+
+        starRocksAssert.alterMvProperties(
+            "alter materialized view test.mv_with_partition set (\"partition_ttl_number\" = \"1\")");
+
+        DynamicPartitionScheduler dynamicPartitionScheduler = GlobalStateMgr.getCurrentState()
+            .getDynamicPartitionScheduler();
+        Database db = GlobalStateMgr.getCurrentState().getDb("test");
+        OlapTable tbl = (OlapTable) db.getTable("mv_with_partition");
+        dynamicPartitionScheduler.registerTtlPartitionTable(db.getId(), tbl.getId());
+        dynamicPartitionScheduler.runOnceForTest();
+        Assert.assertEquals(Sets.newHashSet("p20230413"), versionMap.values().iterator().next().keySet());
+
+        starRocksAssert
+            .useDatabase("test")
+            .dropMaterializedView("mv_with_partition")
+            .dropTable("tbl_with_partition");
     }
 }
